+++ android/src/main/kotlin/com/connect/module/native_view/CameraPlayerFactory.kt
@@ -1,0 +1,17 @@
+package com.connect.module.native_view
+
+import android.app.Activity
+import android.content.Context
+import io.flutter.plugin.common.BinaryMessenger
+import io.flutter.plugin.common.StandardMessageCodec
+import io.flutter.plugin.platform.PlatformView
+import io.flutter.plugin.platform.PlatformViewFactory
+
+class CameraPlayerFactory(
+    private val messenger: BinaryMessenger,
+    private val activity: Activity
+) : PlatformViewFactory(StandardMessageCodec.INSTANCE) {
+    override fun create(context: Context, viewId: Int, args: Any?): PlatformView {
+        return CameraPlayerWidget(context, activity, viewId, messenger, args)
+    }
+}