+++ android/src/main/kotlin/com/connect/utils/AppConstant.kt
@@ -3,5 +3,6 @@ package com.connect.utils
 class AppConstant {
     companion object {
         const val CHANNEL_CAMERA_MULTI = "channelCameraMulti"
+        const val CHANNEL_CAMERA_PLAYER = "channelCameraPlayer"
     }
 }