+++ lib/channel_app.dart
@@ -1,3 +1,4 @@
 class ChannelApp {
   static String channelCameraMulti = "channelCameraMulti";
+  static String viewPlayer = 'channelCameraPlayer';
 }